v1.1.2

+++ front-end-pis/hooks/SessionUtil.js
@@ -1,34 +1,46 @@
 export const save = (key, data) => {
-  sessionStorage.setItem(key, data);
+  if (typeof window !== "undefined") {
+    sessionStorage.setItem(key, data);
+  }
 };
 
 export const get = (key) => {
-  return sessionStorage.getItem(key);
+  if (typeof window !== "undefined") {
+    return sessionStorage.getItem(key);
+  }
 };
 
 export const saveToken = (key) => {
-  sessionStorage.setItem("token", key);
+  if (typeof window !== "undefined") {
+    sessionStorage.setItem("token", key);
+  }
 };
 
 export const getToken = () => {
-  return sessionStorage.getItem("token");
+  if (typeof window !== "undefined") {
+    return sessionStorage.getItem("token");
+  }
 };
 
 export const getUser = () => {
-  let user = sessionStorage.getItem("user");
-
-  user = user ? JSON.parse(user) : null;
-
-  return user;
+  if (typeof window !== "undefined") {
+    let user = sessionStorage.getItem("user");
+    user = user ? JSON.parse(user) : null;
+    return user;
+  }
 };
 
 export const borrarSesion = () => {
-  sessionStorage.clear();
+  if (typeof window !== "undefined") {
+    sessionStorage.clear();
+  }
 };
 
 export const estaSesion = () => {
-  var token = sessionStorage.getItem("token");
-  return token && (token != "undefined" || token != null || token != "null");
+  if (typeof window !== "undefined") {
+    var token = sessionStorage.getItem("token");
+    return token && (token != "undefined" || token != null || token != "null");
+  }
 };
 
 export const getId = () => {
